requirements.txt
aiogram==3.3.0
python-dotenv==1.0.0
aiosqlite==0.19.0
loguru==0.7.2
psutil==5.9.8
aiofiles==23.2.1
cachetools==5.3.2
async-timeout==4.0.3